+++ data/exposed-redisson/src/main/kotlin/io/bluetape4k/exposed/redisson/repository/SuspendedExposedCacheRepository.kt
@@ -1,13 +1,13 @@
 package io.bluetape4k.exposed.redisson.repository
 
+import io.bluetape4k.exposed.dao.HasIdentifier
+import io.bluetape4k.exposed.redisson.map.EntityMapLoader
+import io.bluetape4k.exposed.redisson.map.EntityMapWriter
 import io.bluetape4k.exposed.redisson.map.ExposedEntityMapWriter
-import io.bluetape4k.exposed.redisson.map.ExposedMapLoader
-import io.bluetape4k.exposed.redisson.map.ExposedMapWriter
+import io.bluetape4k.exposed.redisson.map.SuspendedEntityMapLoader
+import io.bluetape4k.exposed.redisson.map.SuspendedEntityMapWriter
 import io.bluetape4k.exposed.redisson.map.SuspendedExposedEntityMapLoader
 import io.bluetape4k.exposed.redisson.map.SuspendedExposedEntityMapWriter
-import io.bluetape4k.exposed.redisson.map.SuspendedExposedMapLoader
-import io.bluetape4k.exposed.redisson.map.SuspendedExposedMapWriter
-import io.bluetape4k.exposed.repository.HasIdentifier
 import io.bluetape4k.redis.redisson.cache.RedisCacheConfig
 import io.bluetape4k.redis.redisson.cache.localCachedMap
 import io.bluetape4k.redis.redisson.cache.mapCache
@@ -96,14 +96,14 @@ abstract class AbstractSuspendedExposedCacheRepository<T: HasIdentifier<ID>, ID:
 ): SuspendedExposedCacheRepository<T, ID> {
 
     /**
-     * DB의 정보를 Read Through로 캐시에 로딩하는 [ExposedMapLoader] 입니다.
+     * DB의 정보를 Read Through로 캐시에 로딩하는 [EntityMapLoader] 입니다.
      */
-    protected open val mapLoaderAsync: SuspendedExposedMapLoader<ID, T> by lazy {
+    protected open val mapLoaderAsync: SuspendedEntityMapLoader<ID, T> by lazy {
         SuspendedExposedEntityMapLoader(entityTable, scope) { toEntity() }
     }
 
     /**
-     * [ExposedMapWriter] 에서 캐시에서 변경된 내용을 Write Through로 DB에 반영하는 함수입니다.
+     * [EntityMapWriter] 에서 캐시에서 변경된 내용을 Write Through로 DB에 반영하는 함수입니다.
      */
     protected open fun doUpdateEntity(statement: UpdateStatement, entity: T) {
         if (config.isReadWrite) {
@@ -112,7 +112,7 @@ abstract class AbstractSuspendedExposedCacheRepository<T: HasIdentifier<ID>, ID:
     }
 
     /**
-     * [ExposedMapWriter] 에서 캐시에서 추가된 내용을 Write Through로 DB에 반영하는 함수입니다.
+     * [EntityMapWriter] 에서 캐시에서 추가된 내용을 Write Through로 DB에 반영하는 함수입니다.
      */
     protected open fun doBatchInsertEntity(statement: BatchInsertStatement, entity: T) {
         if (config.isReadWrite) {
@@ -124,7 +124,7 @@ abstract class AbstractSuspendedExposedCacheRepository<T: HasIdentifier<ID>, ID:
      * Write Through 모드라면 [ExposedEntityMapWriter]를 생성하여 제공합니다.
      * Read Through Only 라면 null을 반환합니다.
      */
-    protected val mapWriterAsync: SuspendedExposedMapWriter<ID, T>? by lazy {
+    protected val mapWriterAsync: SuspendedEntityMapWriter<ID, T>? by lazy {
         when {
             config.isReadWrite -> SuspendedExposedEntityMapWriter(
                 scope = scope,
