+++ data/exposed-redisson/src/test/kotlin/io/bluetape4k/exposed/redisson/repository/scenarios/ReadThroughScenario.kt
@@ -1,9 +1,9 @@
 package io.bluetape4k.exposed.redisson.repository.scenarios
 
 import io.bluetape4k.collections.toVarargArray
+import io.bluetape4k.exposed.dao.HasIdentifier
 import io.bluetape4k.exposed.redisson.AbstractRedissonTest
 import io.bluetape4k.exposed.redisson.repository.ExposedCacheRepository
-import io.bluetape4k.exposed.repository.HasIdentifier
 import io.bluetape4k.exposed.tests.TestDB
 import io.bluetape4k.logging.KLogging
 import org.amshove.kluent.shouldBeEqualTo
