+++ data/exposed/src/main/kotlin/io/bluetape4k/exposed/dao/HasIdentifier.kt
@@ -1,4 +1,4 @@
-package io.bluetape4k.exposed.repository
+package io.bluetape4k.exposed.dao
 
 import java.io.Serializable
